FR3

+++ build.sh
@@ -1,8 +1,7 @@
 #!/bin/bash -e
 . /etc/profile.d/modules.sh
 SOURCE_FILE=${NAME}-${VERSION}.tar.gz
-
-module load ci
+module add ci
 echo ${SOFT_DIR}
 mkdir -p ${SOFT_DIR}
 echo ${WORKSPACE}
@@ -11,16 +10,27 @@ echo ${SRC_DIR}
 mkdir -p ${SRC_DIR}
 echo ${NAME}
 echo ${VERSION}
-#module load gcc/4.8.2
-if [[ ! -s ${SRC_DIR}/${SOURCE_FILE} ]] ; then
+
+if [ ! -e ${SRC_DIR}/${SOURCE_FILE}.lock ] && [ ! -s ${SRC_DIR}/${SOURCE_FILE} ] ; then
+  touch  ${SRC_DIR}/${SOURCE_FILE}.lock
   echo "Seems the file is not available locally, downloading"
   mkdir -p ${SRC_DIR}
 	wget http://zlib.net/${SOURCE_FILE} -O ${SRC_DIR}/${SOURCE_FILE}
+  echo "releasing lock"
+  rm -v ${SRC_DIR}/${SOURCE_FILE}.lock
+elif [ -e ${SRC_DIR}/${SOURCE_FILE}.lock ] ; then
+  # Someone else has the file, wait till it's released
+  while [ -e ${SRC_DIR}/${SOURCE_FILE}.lock ] ; do
+    echo " There seems to be a download currently under way, will check again in 5 sec"
+    sleep 5
+  done
 fi
 
 tar xvzf ${SRC_DIR}/${SOURCE_FILE} -C ${WORKSPACE} --skip-old-files
 ls ${WORKSPACE}
 
 cd ${WORKSPACE}/${NAME}-${VERSION}
-./configure --prefix=${SOFT_DIR}
-make
+./configure \
+--archs="-arch ${ARCH}" \
+--prefix=${SOFT_DIR}
+make -j2
